+++ Scene.h
@@ -12,13 +12,13 @@ public:
   bool readScene(const char *filename);
   bool readSceneValues(stringstream &s, const int numvals, float * values);
   Camera getCamera();
-  float getDepth();
-  float getHeight();
-  float getWidth();
+  int getDepth();
+  int getHeight();
+  int getWidth();
 private:
   Camera sceneCamera;
-  float depth;
-  float height;
-  float width;
+  int depth;
+  int height;
+  int width;
 
 };
